ChartsData - separate classes, make them final
