+++ tests/Data/Charts/Coordinates.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Validator\Tests\Data\Charts;
+
+use Yiisoft\Validator\Rule\Number;
+
+final class Coordinates
+{
+    #[Number(min: -10, max: 10, formatter: new CustomFormatter())]
+    private int $x;
+    #[Number(min: -10, max: 10)]
+    private int $y;
+}
